매입,매수 컨트롤러 추가작업

+++ database/migrations/2019_11_25_104943_create_stock_trades_table.php
@@ -17,10 +17,10 @@ class CreateStockTradesTable extends Migration
             $table->bigIncrements('id');
             $table->bigInteger('user_id')->comment('유저아이디');
             $table->string('title')->comment('타이틀');
-            $table->bigInteger('before_amount')->comment('이전금액');
-            $table->bigInteger('calc_amount')->comment('가감금액');
-            $table->bigInteger('fee_amount')->comment('수수료');
-            $table->bigInteger('now_amount')->comment('현재금액');
+            $table->decimal('before_amount',10,2)->comment('이전금액');
+            $table->decimal('calc_amount',10,2)->comment('가감금액');
+            $table->decimal('fee_amount',10,2)->comment('수수료');
+            $table->decimal('now_amount',10,2)->comment('현재금액');
             $table->timestamp('created_at')->index()->default(DB::raw('CURRENT_TIMESTAMP'))->comment('생성일자');
             $table->timestamp('updated_at')->default(DB::raw('CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP'))->comment('수정일자');
             $table->softDeletes();
